requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
djangorestframework-gis>=1.0
django-cors-headers>=4.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
celery>=5.3.0
django-celery-beat>=2.5.0
redis>=4.5.0
Pillow>=10.0.0
gunicorn>=21.0.0
